Don't need to override noteOn/noteOff in each instrument

+++ src/Tonic/stk_Clarinet.cpp
@@ -31,14 +31,14 @@ namespace Tonic {
     }
 
     void  Clarinet_::computeSynthesisBlock(const SynthesisContext_& context) {
-      unsigned long bufferFrames = (unsigned int)outputFrames_.frames();
+      unsigned long bufferFrames = outputFrames_.frames();
       stk::StkFloat* samples = (stk::StkFloat*)&outputFrames_[0];
       const stk::StkFrames& lastframe = clarinet_.lastFrame();
 
       float frequencyValue = frequency_.tick(context).value;
       //clarinet_.setFrequency(frequencyValue);
     
-      for (unsigned int i = 0; i < bufferFrames; i++) {
+      for (unsigned long i = 0; i < bufferFrames; i++) {
         clarinet_.tick();
         for (int j = 0; j < lastframe.channels(); j++)
           *samples++ = lastframe[j];
