+++ src/Tonic/stk_Rhodey.cpp
@@ -31,14 +31,14 @@ namespace Tonic {
     }
 
     void  Rhodey_::computeSynthesisBlock(const SynthesisContext_& context) {
-      unsigned long bufferFrames = (unsigned int)outputFrames_.frames();
+      unsigned long bufferFrames = outputFrames_.frames();
       stk::StkFloat* samples = (stk::StkFloat*)&outputFrames_[0];
       const stk::StkFrames& lastframe = rhodey_.lastFrame();
 
       float frequencyValue = frequency_.tick(context).value;
       //rhodey_.setFrequency(frequencyValue);
     
-      for (unsigned int i = 0; i < bufferFrames; i++) {
+      for (unsigned long i = 0; i < bufferFrames; i++) {
         rhodey_.tick();
         for (int j = 0; j < lastframe.channels(); j++)
           *samples++ = lastframe[j];
